requirements.txt
PyOpenGL
numpy
pyzed
progressbar2
tifffile
scipy
matplotlib
scikit-image
